chore(desktop): split palette integration into stacked PR
Keep the task query language and saved-search surfaces in this layer. Move command palette integration into a new top layer.

Generated-By: PostHog Desktop
Task-Id: ffe7092e-7014-4915-8595-70fe560eaa20
