+++ products/desktop/packages/shared/src/analytics-events.ts
@@ -57,11 +57,6 @@ export type CommandMenuAction =
   | "open-artifact"
   | "open-channel"
   | "open-command-center"
-  | "save-feed"
-  | "complete-filter"
-  | "show-all-matches"
-  | "repair-query"
-  | "open-feed"
   | "open-inbox"
   | "open-archived"
   | "open-loops"
