+++ products/desktop/packages/ui/src/features/command/CommandMenu.tsx
@@ -22,24 +22,18 @@ import {
   AutocompleteLabel,
   AutocompleteList,
   AutocompleteStatus,
-  cn,
   Dialog,
   DialogContent,
   Kbd,
-  KbdGroup,
 } from "@posthog/quill";
 import { LOOPS_FLAG, PROJECT_BLUEBIRD_FLAG } from "@posthog/shared";
 import {
   ANALYTICS_EVENTS,
   type CommandMenuAction,
 } from "@posthog/shared/analytics-events";
+import type { Task } from "@posthog/shared/domain-types";
 import { useArchivedTaskIds } from "@posthog/ui/features/archive/useArchivedTaskIds";
 import { channelGlyph } from "@posthog/ui/features/canvas/components/channelGlyph";
-import {
-  EDITOR_TEXT_CLASS,
-  FeedQueryHighlight,
-} from "@posthog/ui/features/canvas/components/FeedQueryInput";
-import { TaskFeedModal } from "@posthog/ui/features/canvas/components/TaskFeedModal";
 import { useChannels } from "@posthog/ui/features/canvas/hooks/useChannels";
 import { useChannelsLayout } from "@posthog/ui/features/canvas/hooks/useChannelsLayout";
 import { useTaskChannelMap } from "@posthog/ui/features/canvas/hooks/useTaskChannelMap";
@@ -55,13 +49,7 @@ import {
   formatHotkeyParts,
   SHORTCUTS,
 } from "@posthog/ui/features/command/keyboard-shortcuts";
-import { PaletteFilterChips } from "@posthog/ui/features/command/PaletteFilterChips";
-import { TaskCommandIcon } from "@posthog/ui/features/command/TaskCommandIcon";
 import { taskSearchDelay } from "@posthog/ui/features/command/taskSearchQuery";
-import {
-  matchSummary,
-  useFeedQueryCommands,
-} from "@posthog/ui/features/command/useFeedQueryCommands";
 import { useFileSearchContext } from "@posthog/ui/features/command/useFileSearchContext";
 import {
   type Command,
@@ -76,7 +64,9 @@ import {
   closeSettings,
   openSettings,
 } from "@posthog/ui/features/settings/hooks/useOpenSettings";
+import { TaskIcon } from "@posthog/ui/features/sidebar/components/items/TaskIcon";
 import { useSidebarStore } from "@posthog/ui/features/sidebar/sidebarStore";
+import { useTaskPrStatus } from "@posthog/ui/features/sidebar/useTaskPrStatus";
 import { useTasks } from "@posthog/ui/features/tasks/useTasks";
 import { useWorkspaces } from "@posthog/ui/features/workspace/useWorkspace";
 import { LoopIcon } from "@posthog/ui/primitives/LoopIcon";
@@ -86,7 +76,6 @@ import {
   navigateToArchived,
   navigateToChannel,
   navigateToCommandCenter,
-  navigateToFeed,
   navigateToInbox,
   navigateToLoops,
 } from "@posthog/ui/router/navigationBridge";
@@ -108,102 +97,38 @@ import {
   ZoomInIcon,
   ZoomOutIcon,
 } from "@radix-ui/react-icons";
-import {
-  type KeyboardEvent as ReactKeyboardEvent,
-  useCallback,
-  useEffect,
-  useLayoutEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 
 interface CommandMenuProps {
   open: boolean;
   onOpenChange: (open: boolean) => void;
 }
 
-const DEFAULT_RESULT_LIMIT = 8;
-const COLLAPSED_CHIP_COUNT = 5;
-
-function PaletteQueryMirror({
-  query,
-  wrapRef,
-  visible,
-}: {
-  query: string;
-  wrapRef: React.RefObject<HTMLDivElement | null>;
-  visible: boolean;
-}) {
-  const mirrorRef = useRef<HTMLDivElement>(null);
-  const [box, setBox] = useState<{
-    left: number;
-    top: number;
-    width: number;
-    height: number;
-    padLeft: number;
-    padRight: number;
-  } | null>(null);
-
-  // A child layout effect runs before the parent ref attaches.
-  useEffect(() => {
-    if (!visible) return;
-    const wrap = wrapRef.current;
-    const input = wrap?.querySelector("input");
-    if (!wrap || !input) return;
-    const measure = () => {
-      const w = wrap.getBoundingClientRect();
-      const r = input.getBoundingClientRect();
-      const style = getComputedStyle(input);
-      setBox({
-        left: r.left - w.left,
-        top: r.top - w.top,
-        width: r.width,
-        height: r.height,
-        padLeft: Number.parseFloat(style.paddingLeft),
-        padRight: Number.parseFloat(style.paddingRight),
-      });
-    };
-    measure();
-    const observer = new ResizeObserver(measure);
-    observer.observe(input);
-    const sync = () => {
-      if (mirrorRef.current) mirrorRef.current.scrollLeft = input.scrollLeft;
-    };
-    input.addEventListener("scroll", sync);
-    return () => {
-      observer.disconnect();
-      input.removeEventListener("scroll", sync);
-    };
-  }, [visible, wrapRef]);
-
-  // A completion can change the input scroll position without a scroll event.
-  useLayoutEffect(() => {
-    const input = wrapRef.current?.querySelector("input");
-    if (input && mirrorRef.current)
-      mirrorRef.current.scrollLeft = input.scrollLeft;
+/**
+ * Task icon for the command palette. Renders the same shared `TaskIcon` as
+ * the sidebar — cloud run status, PR/branch status, etc. — deriving its
+ * inputs from the raw task and a per-task PR-status query.
+ */
+function TaskCommandIcon({ task }: { task: Task }) {
+  const { prState, hasDiff } = useTaskPrStatus({
+    id: task.id,
+    cloudPrUrl: null,
+    taskRunEnvironment: task.latest_run?.environment,
   });
-
-  if (!visible || !box) return null;
+  const stateSlackThreadUrl = (
+    task.latest_run?.state as { slack_thread_url?: unknown } | undefined
+  )?.slack_thread_url;
+  const slackThreadUrl =
+    typeof stateSlackThreadUrl === "string" ? stateSlackThreadUrl : undefined;
   return (
-    <div
-      ref={mirrorRef}
-      aria-hidden
-      className={cn(
-        "pointer-events-none absolute z-10 flex items-center overflow-x-hidden",
-        EDITOR_TEXT_CLASS,
-      )}
-      style={{
-        left: box.left,
-        top: box.top,
-        width: box.width,
-        height: box.height,
-        paddingLeft: box.padLeft,
-        paddingRight: box.padRight,
-      }}
-    >
-      <FeedQueryHighlight query={query} />
-    </div>
+    <TaskIcon
+      workspaceMode={task.latest_run?.environment}
+      taskRunStatus={task.latest_run?.status}
+      originProduct={task.origin_product}
+      slackThreadUrl={slackThreadUrl}
+      prState={prState}
+      hasDiff={hasDiff}
+    />
   );
 }
 
@@ -235,12 +160,6 @@ export function CommandMenu({ open, onOpenChange }: CommandMenuProps) {
   );
   const [query, setQuery] = useState("");
   const [remoteQuery, setRemoteQuery] = useState("");
-  // The legacy title search only ever surfaces while the palette is browsing
-  // (see `showRemoteSearch` below). The feed-query `mode` that decides that is
-  // derived further down, after this debounce, so the effect reads the latest
-  // value through a ref instead because a filter query must not start a search whose
-  // results the palette then throws away.
-  const remoteSearchAllowedRef = useRef(true);
   const { repoPath } = useFileSearchContext();
   const canSearchFiles = !!repoPath;
   const openFilePicker = useFileSearchStore((state) => state.openPicker);
@@ -255,10 +174,7 @@ export function CommandMenu({ open, onOpenChange }: CommandMenuProps) {
       setRemoteQuery("");
       return;
     }
-    const timer = window.setTimeout(
-      () => setRemoteQuery(remoteSearchAllowedRef.current ? trimmed : ""),
-      delay,
-    );
+    const timer = window.setTimeout(() => setRemoteQuery(trimmed), delay);
     return () => window.clearTimeout(timer);
   }, [open, query]);
 
@@ -429,18 +345,7 @@ export function CommandMenu({ open, onOpenChange }: CommandMenuProps) {
     ];
 
     const actions: Command[] = [
-      {
-        id: "new-task",
-        label: "New task",
-        keywords: "create",
-        icon: <FileTextIcon className="h-3 w-3 text-gray-11" />,
-        action: "new-task",
-        shortcut: SHORTCUTS.NEW_TASK,
-        onRun: () => {
-          closeSettingsDialog();
-          openTaskInput();
-        },
-      },
+      ...themeOptions,
       {
         id: "toggle-left-sidebar",
         label: "Toggle left sidebar",
@@ -461,8 +366,18 @@ export function CommandMenu({ open, onOpenChange }: CommandMenuProps) {
             },
           ]
         : []),
-      // Last, because the first row is what a stray Enter runs.
-      ...themeOptions,
+      {
+        id: "new-task",
+        label: "New task",
+        keywords: "create",
+        icon: <FileTextIcon className="h-3 w-3 text-gray-11" />,
+        action: "new-task",
+        shortcut: SHORTCUTS.NEW_TASK,
+        onRun: () => {
+          closeSettingsDialog();
+          openTaskInput();
+        },
+      },
     ];
 
     if (canSearchFiles) {
@@ -664,109 +579,25 @@ export function CommandMenu({ open, onOpenChange }: CommandMenuProps) {
     spacesLayout,
   });
 
-  const [feedModalQuery, setFeedModalQuery] = useState<string | null>(null);
-  const onSaveAsFeed = useCallback(
-    (feedQuery: string) => {
-      onOpenChange(false);
-      setFeedModalQuery(feedQuery);
-    },
-    [onOpenChange],
-  );
-
-  const [caret, setCaret] = useState(0);
-  const inputWrapRef = useRef<HTMLDivElement>(null);
-  const trackCaret = useCallback(() => {
-    const position =
-      inputWrapRef.current?.querySelector("input")?.selectionStart;
-    if (position != null) setCaret(position);
-  }, []);
-  // Restore the caret after React applies a completed suggestion.
-  const pendingCaret = useRef<number | null>(null);
-  useLayoutEffect(() => {
-    if (pendingCaret.current == null) return;
-    const input = inputWrapRef.current?.querySelector("input");
-    input?.setSelectionRange(pendingCaret.current, pendingCaret.current);
-    input?.focus();
-    pendingCaret.current = null;
-  });
-  const onApplyFilter = useCallback((next: string, nextCaret: number) => {
-    setQuery(next);
-    setCaret(nextCaret);
-    pendingCaret.current = nextCaret;
-  }, []);
-
-  const [showAllFor, setShowAllFor] = useState<string | null>(null);
-  const resultLimit =
-    showAllFor === query ? Number.MAX_SAFE_INTEGER : DEFAULT_RESULT_LIMIT;
-  const showAllMatches = useCallback(() => setShowAllFor(query), [query]);
-
-  const feedQuery = useFeedQueryCommands({
-    query,
-    caret,
-    enabled: open && spacesLayout,
-    limit: resultLimit,
-    onApply: onApplyFilter,
-    onShowAll: showAllMatches,
-  });
-  const {
-    mode,
-    scope,
-    hasFilterTokens,
-    searchText,
-    keyChips,
-    matchCount,
-    partialResults,
-    shownCount,
-    hasRepairs,
-  } = feedQuery;
-
-  // Keep the debounce's gate in step with the condition that decides whether
-  // the remote search is even rendered, so the two can't drift.
-  useEffect(() => {
-    const browsing = mode === "browsing" || mode === "completingKey";
-    remoteSearchAllowedRef.current = browsing && !scope;
-  }, [mode, scope]);
-
-  const sections = useMemo(() => {
-    const browsing = mode === "browsing" || mode === "completingKey";
-    const showCommands = browsing && (!scope || scope === "command");
-    const showChannels = browsing && (!scope || scope === "space");
-    const showPlainTasks = browsing && !scope;
-    const showRemoteSearch = browsing && !scope;
-    return prioritizeExactCommandMatches(
-      [
-        ...feedQuery.sections,
-        ...(showRemoteSearch ? searchSections : []),
-        ...(showCommands ? commandSections : []),
-        ...(showChannels ? channelSections : []),
-        ...(showPlainTasks ? taskSections : []),
-      ],
-      searchText,
-    );
-  }, [
-    feedQuery.sections,
-    searchSections,
-    commandSections,
-    channelSections,
-    taskSections,
-    mode,
-    scope,
-    searchText,
-  ]);
-
-  const paletteFilter = useCallback(
-    (command: { label: string; keywords?: string }) =>
-      matchesCommandSearch(command, searchText),
-    [searchText],
+  // Commands, channels, and tasks share a single filterable list.
+  const sections = useMemo(
+    () =>
+      prioritizeExactCommandMatches(
+        [
+          ...searchSections,
+          ...commandSections,
+          ...channelSections,
+          ...taskSections,
+        ],
+        query,
+      ),
+    [searchSections, commandSections, channelSections, taskSections, query],
   );
 
   const allCommands = useMemo(
     () => sections.flatMap((s) => s.items),
     [sections],
   );
-
-  const highlightedId = useRef<string | null>(null);
-  const showMatchSummary = mode === "querying" || matchCount != null;
 
   const handleSelect = (id: string | null): void => {
     if (id === null) return;
@@ -777,185 +608,91 @@ export function CommandMenu({ open, onOpenChange }: CommandMenuProps) {
       channel_id: cmd.channelId,
     });
     cmd.onRun();
-    if (cmd.keepOpen) return;
     onOpenChange(false);
     setQuery("");
   };
 
-  const onInputKeyDown = (event: ReactKeyboardEvent<HTMLInputElement>) => {
-    if (
-      event.key.toLowerCase() === "s" &&
-      (event.metaKey || event.ctrlKey) &&
-      hasFilterTokens
-    ) {
-      event.preventDefault();
-      track(ANALYTICS_EVENTS.COMMAND_MENU_ACTION, { action_type: "save-feed" });
-      onSaveAsFeed(query.trim());
-      return;
-    }
-    if (event.key !== "Tab") return;
-    if (mode === "completingKey" && keyChips.length > 0) {
-      event.preventDefault();
-      keyChips[0].apply();
-      return;
-    }
-    const highlighted = allCommands.find(
-      (c) => c.id === highlightedId.current && c.keepOpen,
-    );
-    const target = highlighted ?? allCommands.find((c) => c.keepOpen);
-    if (!target) return;
-    event.preventDefault();
-    target.onRun();
-  };
-
   return (
-    <>
-      <Dialog open={open} onOpenChange={onOpenChange}>
-        <DialogContent
-          className="w-[720px] max-w-[90vw] gap-0 p-0"
-          showCloseButton={false}
+    <Dialog open={open} onOpenChange={onOpenChange}>
+      <DialogContent
+        className="w-[720px] max-w-[90vw] gap-0 p-0"
+        showCloseButton={false}
+      >
+        <Autocomplete<Command>
+          inline
+          defaultOpen
+          items={sections}
+          value={query}
+          autoHighlight="always"
+          keepHighlight
+          onValueChange={(val, eventDetails) => {
+            if (eventDetails.reason !== "input-change") return;
+            if (typeof val === "string") {
+              setQuery(val);
+            }
+          }}
+          filter={matchesCommandSearch}
         >
-          <Autocomplete<Command>
-            inline
-            defaultOpen
-            items={sections}
-            value={query}
-            autoHighlight
-            keepHighlight
-            onItemHighlighted={(value) => {
-              highlightedId.current = typeof value === "string" ? value : null;
-            }}
-            onValueChange={(val, eventDetails) => {
-              if (typeof val !== "string") return;
-              if (eventDetails.reason === "input-change") {
-                setQuery(val);
-                trackCaret();
-                return;
-              }
-              if (val === "") setQuery("");
-            }}
-            filter={paletteFilter}
-          >
-            <div ref={inputWrapRef} className="relative">
-              <PaletteQueryMirror
-                query={query}
-                wrapRef={inputWrapRef}
-                visible={spacesLayout}
-              />
-              <AutocompleteInput
-                placeholder={
-                  spacesLayout
-                    ? "Search commands and tasks, or filter with created-by:"
-                    : bluebirdEnabled
-                      ? "Search commands, channels, and tasks…"
-                      : "Search commands and tasks…"
-                }
-                autoFocus
-                showClear
-                className={
-                  spacesLayout
-                    ? cn(
-                        "[&_input]:whitespace-pre [&_input]:font-mono [&_input]:text-[13px] [&_input]:tracking-normal",
-                        "[&_input]:text-transparent [&_input]:caret-(--gray-12) [&_input]:placeholder:text-(--gray-9)",
-                        "[&_input]:selection:bg-(--blue-a4) [&_input]:selection:text-transparent",
-                      )
-                    : undefined
-                }
-                onKeyDown={onInputKeyDown}
-                onKeyUp={trackCaret}
-                onClick={trackCaret}
-                onSelect={trackCaret}
-              />
-            </div>
-            {keyChips.length > 0 && (
-              <PaletteFilterChips
-                chips={keyChips}
-                collapsedCount={COLLAPSED_CHIP_COUNT}
-                filtering={mode === "completingKey"}
-              />
-            )}
-            {showMatchSummary ? (
-              <div className="border-(--gray-a4) border-b px-3 py-1.5 text-(--gray-9) text-xs tabular-nums">
-                {partialResults
-                  ? "Some matching tasks may not be shown."
-                  : matchSummary(matchCount, shownCount, hasRepairs)}
-              </div>
-            ) : (
-              query !== "" && (
-                <AutocompleteStatus
-                  emptyContent={
-                    <span>
-                      No results for <strong>"{query}"</strong>
-                    </span>
-                  }
-                />
-              )
-            )}
-            <AutocompleteList className="max-h-[60vh]">
-              {(section: CommandSection) => (
-                <AutocompleteGroup key={section.label} items={section.items}>
-                  <AutocompleteLabel>{section.label}</AutocompleteLabel>
-                  <AutocompleteCollection>
-                    {(cmd: Command) => (
-                      <AutocompleteItem
-                        key={cmd.id}
-                        value={cmd.id}
-                        onClick={() => handleSelect(cmd.id)}
-                        className="group flex h-auto! min-h-7 w-full items-center gap-2 py-1.5 pr-2 text-left [&>span]:w-full [&>span]:overflow-visible"
-                      >
-                        {cmd.icon}
-                        <span className="wrap-break-word min-w-0 whitespace-normal">
-                          {cmd.label}
-                          {cmd.subtitle && (
-                            <span className="block truncate text-gray-9 text-xs">
-                              {cmd.subtitle}
-                            </span>
-                          )}
+          <AutocompleteInput
+            placeholder={
+              bluebirdEnabled
+                ? `Search commands, ${spacesLayout ? "spaces" : "channels"}, and tasks…`
+                : "Search commands and tasks…"
+            }
+            autoFocus
+            showClear
+          />
+          <AutocompleteStatus
+            emptyContent={
+              <span>
+                No results for <strong>"{query}"</strong>
+              </span>
+            }
+          />
+          <AutocompleteList className="max-h-[60vh]">
+            {(section: CommandSection) => (
+              <AutocompleteGroup key={section.label} items={section.items}>
+                <AutocompleteLabel>{section.label}</AutocompleteLabel>
+                <AutocompleteCollection>
+                  {(cmd: Command) => (
+                    <AutocompleteItem
+                      key={cmd.id}
+                      value={cmd.id}
+                      onClick={() => handleSelect(cmd.id)}
+                      // Long task names wrap instead of truncating, so the
+                      // item must grow: min-height, not a fixed height. Quill
+                      // wraps our children in an inner content span; force it to
+                      // fill the row (so a trailing shortcut can `ml-auto` to the
+                      // end) and let it overflow visibly so the shortcut Kbd
+                      // boxes aren't clipped by the wrapper's `truncate`.
+                      className="group flex h-auto! min-h-7 w-full items-center gap-2 py-1.5 pr-2 text-left [&>span]:w-full [&>span]:overflow-visible"
+                    >
+                      {cmd.icon}
+                      <span className="wrap-break-word min-w-0 whitespace-normal">
+                        {cmd.label}
+                      </span>
+                      {cmd.detail && (
+                        <span className="shrink-0 text-gray-9">
+                          · {cmd.detailPrefix ?? "#"}
+                          {cmd.detail}
                         </span>
-                        {cmd.detail && (
-                          <span className="shrink-0 text-gray-9">
-                            · {cmd.detailPrefix ?? "#"}
-                            {cmd.detail}
-                          </span>
-                        )}
-                        {cmd.shortcut && (
-                          <span className="ml-auto flex shrink-0 items-center gap-2 pl-2">
-                            {formatHotkeyParts(cmd.shortcut).map((part) => (
-                              <Kbd key={part}>{part}</Kbd>
-                            ))}
-                          </span>
-                        )}
-                      </AutocompleteItem>
-                    )}
-                  </AutocompleteCollection>
-                </AutocompleteGroup>
-              )}
-            </AutocompleteList>
-          </Autocomplete>
-          <CommandKeyHints>
-            {hasFilterTokens && (
-              <div className="flex items-center gap-2">
-                <KbdGroup>
-                  <Kbd>{formatHotkeyParts("mod+s").join("")}</Kbd>
-                </KbdGroup>
-                <span className="text-xs">
-                  save search
-                  {matchCount === 0 && " · no matches yet"}
-                </span>
-              </div>
+                      )}
+                      {cmd.shortcut && (
+                        <span className="ml-auto flex shrink-0 items-center gap-2 pl-2">
+                          {formatHotkeyParts(cmd.shortcut).map((part) => (
+                            <Kbd key={part}>{part}</Kbd>
+                          ))}
+                        </span>
+                      )}
+                    </AutocompleteItem>
+                  )}
+                </AutocompleteCollection>
+              </AutocompleteGroup>
             )}
-          </CommandKeyHints>
-        </DialogContent>
-      </Dialog>
-      <TaskFeedModal
-        open={feedModalQuery !== null}
-        onOpenChange={(modalOpen) => {
-          if (!modalOpen) setFeedModalQuery(null);
-        }}
-        initialQuery={feedModalQuery ?? undefined}
-        surface="command_menu"
-        onCreated={(feed) => navigateToFeed(feed.id)}
-      />
-    </>
+          </AutocompleteList>
+        </Autocomplete>
+        <CommandKeyHints />
+      </DialogContent>
+    </Dialog>
   );
 }
